AE_Cabinet, Combfilter, Soundgenerator: resolved type issues (signed/unsigned integers) and replaced functions (pow) for correct type handling

+++ examples/c15_audio_engine/ae_soundgenerator.cpp
@@ -44,8 +44,8 @@ void ae_soundgenerator::init(float _samplerate, uint32_t _vn)
     m_chiB_a0 = 0.f;
     m_chiB_a1 = 0.f;
 
-    m_OscA_randVal_int = _vn + 1;
-    m_OscB_randVal_int = _vn + 1 + 111;
+    m_OscA_randVal_int = static_cast<int32_t>(_vn) + 1;
+    m_OscB_randVal_int = static_cast<int32_t>(_vn) + 1 + 111;
 
 }
